binance test cases finished (untested)

+++ structs/structs.go
@@ -24,8 +24,8 @@ type TradeData struct {
 	TimeStamp uint
 	Date      uint
 	Symbol    string
-	BidPrice  uint
-	BidSize   uint
-	AskPrice  uint
-	AskSize   uint
+	BidPrice  float32
+	BidSize   float32
+	AskPrice  float32
+	AskSize   float32
 }
